+++ unit_tests/test_string.py
@@ -1,8 +1,18 @@
+from __future__ import division
+
+import itertools
+import math
 import unittest
+from collections import Counter, OrderedDict
 
 import pytest
+import six
+from future.standard_library import install_aliases
+from six.moves import map, zip
 
 from boofuzz import *
+
+install_aliases()
 
 
 @pytest.fixture(autouse=True)
@@ -18,15 +28,21 @@ class TestString(unittest.TestCase):
         self.default_default_value = "\x00" * len(self.default_value)
         return String(name="boofuzz-unit-test-name", default_value=self.default_value)
 
-    def _given_bytes_max_len(self, max_len):
-        self.default_value = b"ABCDEFGH"
+    def _given_string_max_len(self, max_len):
+        self.default_value = "ABCDEFGH"
         self.default_default_value = "\x00" * len(self.default_value)
-        return Bytes(name="boofuzz-unit-test-name", default_value=self.default_value, max_len=max_len)
+        return String(name="boofuzz-unit-test-name", default_value=self.default_value, max_len=max_len)
 
-    def _given_bytes_size(self, size, padding):
-        self.default_value = b"ABCDEFGH"
+    def _given_string_size(self, size, padding, encoding):
+        self.default_value = "ABCDEFGH"
         self.default_default_value = "\x00" * len(self.default_value)
-        return Bytes(name="boofuzz-unit-test-name", default_value=self.default_value, size=size, padding=padding)
+        return String(
+            name="boofuzz-unit-test-name",
+            default_value=self.default_value,
+            size=size,
+            padding=padding,
+            encoding=encoding,
+        )
 
     def test_mutations(self):
         uut = self._given_string()
@@ -43,113 +59,156 @@ class TestString(unittest.TestCase):
             self.assertEqual(self.default_default_value * expected, actual)
 
         for sequence in String.long_string_seeds:
-            for size in String._long_string_lengths:
-                for delta in String._long_string_deltas:
-                    n += 1
-                    self.assertEqual(sequence * (size + delta), next(generator))
+            for size in [
+                length + delta
+                for length, delta in itertools.product(String._long_string_lengths, String._long_string_deltas)
+            ]:
+                n += 1
+                expected = sequence * math.ceil(size / len(sequence))
+                self.assertEqual(expected[:size], next(generator))
+
+            for size in String._extra_long_string_lengths:
+                n += 1
+                expected = sequence * math.ceil(size / len(sequence))
+                self.assertEqual(expected[:size], next(generator))
 
         for size in String._long_string_lengths:
             s = "D" * size
             for loc in uut.random_indices[size]:
-                s = s[:loc] + "\x00" + s[loc:]
                 n += 1
-                self.assertEqual(s, next(generator))
-
-        for sequence in String.long_string_seeds:
-            for size in String._extra_long_string_lengths:
-                n += 1
-                self.assertEqual(sequence * size, next(generator))
+                expected = s[:loc] + "\x00" + s[loc + 1 :]
+                self.assertEqual(expected, next(generator))
 
         self.assertRaises(StopIteration, lambda: next(generator))
         self.assertEqual(n, uut.num_mutations(default_value=self.default_value))
+        list_of_duplicates = [
+            item for item, count in Counter(uut.mutations(default_value=self.default_value)).items() if count > 1
+        ]
+        self.assertEqual(0, len(list_of_duplicates))
 
-    # def test_mutations_max_len(self):
-    #     max_len = 7
-    #     uut = self._given_bytes_max_len(max_len=max_len)
-    #     generator = uut.mutations(default_value=self.default_default_value)
-    #
-    #     def truncate(b):
-    #         return b[0:max_len]
-    #
-    #     n = 0
-    #     for expected, actual in zip(map(truncate, uut._fuzz_library), generator):
-    #         n += 1
-    #         self.assertEqual(expected, actual)
-    #     for expected, actual in zip(uut._mutators_of_default_value, generator):
-    #         n += 1
-    #         self.assertEqual(truncate(expected(self.default_value)), actual(self.default_value))
-    #     for expected, actual in zip(map(truncate, uut._magic_debug_values), generator):
-    #         n += 1
-    #         self.assertEqual(expected, actual)
-    #     for i in range(0, len(self.default_default_value)):
-    #         for expected_fuzz_string in uut._fuzz_strings_1byte:
-    #             n += 1
-    #             actual_callable = next(generator)
-    #             actual_value = actual_callable(self.default_value)
-    #             expected_value = self.default_value[0:i] + expected_fuzz_string + self.default_value[i + 1 :]
-    #             self.assertEqual(truncate(expected_value), actual_value)
-    #     for i in range(0, len(self.default_default_value) - 1):
-    #         for expected_fuzz_string in uut._fuzz_strings_2byte:
-    #             n += 1
-    #             actual_callable = next(generator)
-    #             actual_value = actual_callable(self.default_value)
-    #             expected_value = self.default_value[0:i] + expected_fuzz_string + self.default_value[i + 2 :]
-    #             self.assertEqual(truncate(expected_value), actual_value)
-    #     for i in range(0, len(self.default_default_value) - 3):
-    #         for expected_fuzz_string in uut._fuzz_strings_4byte:
-    #             n += 1
-    #             actual_callable = next(generator)
-    #             actual_value = actual_callable(self.default_value)
-    #             expected_value = self.default_value[0:i] + expected_fuzz_string + self.default_value[i + 4 :]
-    #             self.assertEqual(truncate(expected_value), actual_value)
-    #
-    #     self.assertEqual(n, uut.num_mutations(default_value=self.default_value))
-    #
-    # def test_mutations_size(self):
-    #     size = 5
-    #     pad = b"\x41"
-    #
-    #     def fit_to_size(b):
-    #         fullpad = b"\x41" * 5
-    #         pad_len = max(0, size - len(b))
-    #         return b[0:size] + fullpad[0:pad_len]
-    #
-    #     uut = self._given_bytes_size(size=size, padding=pad)
-    #     generator = uut.mutations(default_value=self.default_default_value)
-    #
-    #     n = 0
-    #     for expected, actual in zip(map(fit_to_size, uut._fuzz_library), generator):
-    #         n += 1
-    #         self.assertEqual(expected, actual)
-    #     for expected, actual in zip(uut._mutators_of_default_value, generator):
-    #         n += 1
-    #         self.assertEqual(fit_to_size(expected(self.default_value)), actual(self.default_value))
-    #     for expected, actual in zip(map(fit_to_size, uut._magic_debug_values), generator):
-    #         n += 1
-    #         self.assertEqual(expected, actual)
-    #     for i in range(0, len(self.default_default_value)):
-    #         for expected_fuzz_string in uut._fuzz_strings_1byte:
-    #             n += 1
-    #             actual_callable = next(generator)
-    #             actual_value = actual_callable(self.default_value)
-    #             expected_value = self.default_value[0:i] + expected_fuzz_string + self.default_value[i + 1 :]
-    #             self.assertEqual(fit_to_size(expected_value), actual_value)
-    #     for i in range(0, len(self.default_default_value) - 1):
-    #         for expected_fuzz_string in uut._fuzz_strings_2byte:
-    #             n += 1
-    #             actual_callable = next(generator)
-    #             actual_value = actual_callable(self.default_value)
-    #             expected_value = self.default_value[0:i] + expected_fuzz_string + self.default_value[i + 2 :]
-    #             self.assertEqual(fit_to_size(expected_value), actual_value)
-    #     for i in range(0, len(self.default_default_value) - 3):
-    #         for expected_fuzz_string in uut._fuzz_strings_4byte:
-    #             n += 1
-    #             actual_callable = next(generator)
-    #             actual_value = actual_callable(self.default_value)
-    #             expected_value = self.default_value[0:i] + expected_fuzz_string + self.default_value[i + 4 :]
-    #             self.assertEqual(fit_to_size(expected_value), actual_value)
-    #
-    #     self.assertEqual(n, uut.num_mutations(default_value=self.default_value))
+    def test_mutations_max_len(self):
+        lengths = [5, 10, 128, 1000, 100000]
+
+        for max_len in lengths:
+            uut = self._given_string_max_len(max_len=max_len)
+            generator = uut.mutations(default_value=self.default_default_value)
+
+            def truncate(b):
+                return b[:max_len]
+
+            n = 0
+            for expected, actual in zip(OrderedDict.fromkeys(list(map(truncate, String._fuzz_library))), generator):
+                n += 1
+                self.assertEqual(expected, actual)
+
+            for expected, actual in zip(String._variable_mutation_multipliers, generator):
+                n += 1
+                self.assertEqual(truncate(self.default_default_value * expected), actual)
+                if max_len <= len(self.default_default_value * expected):
+                    break
+
+            for sequence in String.long_string_seeds:
+                for size in [
+                    length + delta
+                    for length, delta in itertools.product(String._long_string_lengths, String._long_string_deltas)
+                ]:
+                    if size <= max_len:
+                        n += 1
+                        expected = sequence * math.ceil(size / len(sequence))
+                        self.assertEqual(truncate(expected[:size]), next(generator))
+
+                for size in String._extra_long_string_lengths:
+                    if size <= max_len:
+                        n += 1
+                        expected = sequence * math.ceil(size / len(sequence))
+                        self.assertEqual(truncate(expected[:size]), next(generator))
+
+                if max_len not in String._extra_long_string_lengths + [
+                    length + delta
+                    for length, delta in itertools.product(String._long_string_lengths, String._long_string_deltas)
+                ]:
+                    n += 1
+                    expected = sequence * math.ceil(max_len / len(sequence))
+                    self.assertEqual(truncate(expected), next(generator))
+
+            for size in String._long_string_lengths:
+                if size <= max_len:
+                    s = "D" * size
+                    for loc in uut.random_indices[size]:
+                        expected = s[:loc] + "\x00" + s[loc + 1 :]
+                        n += 1
+                        self.assertEqual(truncate(expected), next(generator))
+
+            self.assertRaises(StopIteration, lambda: next(generator))
+            self.assertEqual(n, uut.num_mutations(default_value=self.default_value))
+            list_of_duplicates = [
+                item for item, count in Counter(uut.mutations(default_value=self.default_value)).items() if count > 1
+            ]
+            self.assertEqual(0, len(list_of_duplicates))
+
+    def test_mutations_size(self):
+        lengths = [5, 10, 128, 1000, 100000]
+        pad = b"\x41"
+        encoding = "utf-8"
+
+        def fit_to_size(b):
+            b = six.ensure_binary(b[:max_len], encoding=encoding)
+            pad_len = max(0, max_len - len(b))
+            return b + pad * pad_len
+
+        for max_len in lengths:
+            uut = self._given_string_size(size=max_len, padding=pad, encoding=encoding)
+            generator = uut.mutations(default_value=self.default_default_value)
+
+            n = 0
+            for expected, actual in zip(OrderedDict.fromkeys(list(map(fit_to_size, String._fuzz_library))), generator):
+                n += 1
+                self.assertEqual(expected, uut.encode(actual))
+
+            for expected, actual in zip(String._variable_mutation_multipliers, generator):
+                n += 1
+                self.assertEqual(fit_to_size(self.default_default_value * expected), uut.encode(actual))
+                if max_len <= len(self.default_default_value * expected):
+                    break
+
+            for sequence in String.long_string_seeds:
+                for size in [
+                    length + delta
+                    for length, delta in itertools.product(String._long_string_lengths, String._long_string_deltas)
+                ]:
+                    if size <= max_len:
+                        n += 1
+                        expected = sequence * math.ceil(size / len(sequence))
+                        self.assertEqual(fit_to_size(expected[:size]), uut.encode(next(generator)))
+
+                for size in String._extra_long_string_lengths:
+                    if size <= max_len:
+                        n += 1
+                        expected = sequence * math.ceil(size / len(sequence))
+                        self.assertEqual(fit_to_size(expected[:size]), uut.encode(next(generator)))
+
+                if max_len not in String._extra_long_string_lengths + [
+                    length + delta
+                    for length, delta in itertools.product(String._long_string_lengths, String._long_string_deltas)
+                ]:
+                    n += 1
+                    expected = sequence * math.ceil(max_len / len(sequence))
+                    self.assertEqual(fit_to_size(expected), uut.encode(next(generator)))
+
+            for length in String._long_string_lengths:
+                if length <= max_len:
+                    s = "D" * length
+                    for loc in uut.random_indices[length]:
+                        expected = s[:loc] + "\x00" + s[loc + 1 :]
+                        n += 1
+                        self.assertEqual(fit_to_size(expected), uut.encode(next(generator)))
+
+            self.assertRaises(StopIteration, lambda: next(generator))
+            self.assertEqual(n, uut.num_mutations(default_value=self.default_value))
+            list_of_duplicates = [
+                item for item, count in Counter(uut.mutations(default_value=self.default_value)).items() if count > 1
+            ]
+            self.assertEqual(0, len(list_of_duplicates))
 
 
 if __name__ == "__main__":